Split the code into a dart_repl and an embedded dart_repl_sandbox package. Use SendPort/ReceivePort to communicate between the SandboxIsolate and the main isolate.
dart_repl_sandbox has no package dependencies whatsoever. This reduces the chance of conflict between dart_repl and other packages.
Using SendPort/ReceivePort communication allows us to keep the SandboxIsolate very simple. It is mostly a container for values and an environment to evaluate expressions in. All the VM Service code is running the main isolate.